requirements.txt
streamlit
openai
opencv-python-headless
Pillow
matplotlib
numpy
reportlab
scikit-image
